requirements.txt
# Core Bot Framework
python-telegram-bot==21.5
asyncio

# Database
sqlalchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Configuration Management
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging and Monitoring
structlog==23.2.0
colorama==0.4.6

# Utilities
redis==5.0.1
python-dateutil==2.8.2
uuid==1.30

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Production
gunicorn==21.2.0
uvicorn==0.24.0 